refactor(arch): Phase 5 — PR #486 리뷰 2차 수정
- B1: entities/session barrel에서 bcrypt export 제거 (session 책임 범위 외)
- B2: deleteAccountAction deep import → barrel import 통일 + 중복 병합
- S3: entities/session/actions.ts에서 CleanupUnauthorizedError/Result 제거
