+++ src/entities/session/actions.ts
@@ -1,8 +1,4 @@
 'use server';
 
 export { currentUserAction } from './actions/currentUserAction';
-export {
-    cleanupExpiredSessionsAction,
-    CleanupUnauthorizedError,
-} from './actions/cleanupExpiredSessionsAction';
-export type { CleanupExpiredSessionsResult } from './actions/cleanupExpiredSessionsAction';
+export { cleanupExpiredSessionsAction } from './actions/cleanupExpiredSessionsAction';
