+++ src/entities/session/index.ts
@@ -32,4 +32,3 @@ export type {
     PasswordHasher,
     PasswordVerifier,
 } from './lib/types';
-export { bcryptPasswordHasher, bcryptPasswordVerifier } from './lib/bcrypt';
